# feat: improved bytesToUuid func