--- a/examples/benchmark/benchmark.js
+++ b/examples/benchmark/benchmark.js
@@ -9,7 +9,12 @@ const uuidv5 = (typeof window !== 'undefined' && window.uuidv5) || require('uuid
 console.log('Starting. Tests take ~1 minute to run ...');
 
 const array = new Array(16);
-const suite = new Benchmark.Suite();
+
+const suite = new Benchmark.Suite({
+  onError(event) {
+    console.error(event.target.error);
+  },
+});
 
 suite
   .add('uuidv1()', function () {