@ctavan 2020-05-03
Great optimization, thank you!

Interestingly the string concatenation seems to be faster on V8 (Node.js/Chrome), but slower on JavaScriptCore (Safari) and SpiderMonkey (Firefox).

Since this is definitely a perf improvement for Node.js, I will still merge it since performance should only ever be a real concern in serverside use cases, not in the browser.

For the record:

### Safari 13.1

Master:

```
[Log] Starting. Tests take ~1 minute to run ... (benchmark.js, line 9)
[Log] uuidv1() x 373,296 ops/sec ±2.65% (49 runs sampled) (benchmark.js, line 34)
[Log] uuidv1() fill existing array x 589,196 ops/sec ±4.61% (49 runs sampled) (benchmark.js, line 34)
[Log] uuidv4() x 839,177 ops/sec ±1.15% (62 runs sampled) (benchmark.js, line 34)
[Log] uuidv4() fill existing array x 1,366,066 ops/sec ±1.40% (57 runs sampled) (benchmark.js, line 34)
[Log] uuidv3() x 22,352 ops/sec ±0.81% (59 runs sampled) (benchmark.js, line 34)
[Log] uuidv5() x 13,163 ops/sec ±2.27% (56 runs sampled) (benchmark.js, line 34)
[Log] Fastest is uuidv4() fill existing array (benchmark.js, line 37)
```

Branch:

```
[Log] Starting. Tests take ~1 minute to run ... (benchmark.js, line 9)
[Log] uuidv1() x 316,734 ops/sec ±4.59% (47 runs sampled) (benchmark.js, line 39)
[Log] uuidv1() fill existing array x 516,869 ops/sec ±3.99% (45 runs sampled) (benchmark.js, line 39)
[Log] uuidv4() x 653,217 ops/sec ±2.61% (56 runs sampled) (benchmark.js, line 39)
[Log] uuidv4() fill existing array x 1,218,314 ops/sec ±1.52% (55 runs sampled) (benchmark.js, line 39)
[Log] uuidv3() x 14,437 ops/sec ±1.20% (52 runs sampled) (benchmark.js, line 39)
[Log] uuidv5() x 10,993 ops/sec ±2.21% (49 runs sampled) (benchmark.js, line 39)
[Log] Fastest is uuidv4() fill existing array (benchmark.js, line 42)
```

### Firefox 75

Master:

```
Starting. Tests take ~1 minute to run ... benchmark.js:9:9
uuidv1() x 858,098 ops/sec ±2.46% (54 runs sampled) benchmark.js:34:13
uuidv1() fill existing array x 3,341,471 ops/sec ±1.99% (59 runs sampled) benchmark.js:34:13
uuidv4() x 555,726 ops/sec ±1.51% (57 runs sampled) benchmark.js:34:13
uuidv4() fill existing array x 1,007,543 ops/sec ±1.13% (61 runs sampled) benchmark.js:34:13
uuidv3() x 49,755 ops/sec ±3.55% (31 runs sampled) benchmark.js:34:13
uuidv5() x 33,969 ops/sec ±1.27% (57 runs sampled) benchmark.js:34:13
Fastest is uuidv1() fill existing array benchmark.js:37:13
```

This Branch:

```
Starting. Tests take ~1 minute to run ... benchmark.js:9:9
uuidv1() x 284,785 ops/sec ±1.38% (56 runs sampled) benchmark.js:39:13
uuidv1() fill existing array x 2,814,427 ops/sec ±2.47% (54 runs sampled) benchmark.js:39:13
uuidv4() x 275,961 ops/sec ±1.89% (55 runs sampled) benchmark.js:39:13
uuidv4() fill existing array x 1,037,837 ops/sec ±0.44% (61 runs sampled) benchmark.js:39:13
uuidv3() x 26,069 ops/sec ±1.21% (22 runs sampled) benchmark.js:39:13
uuidv5() x 24,065 ops/sec ±2.95% (52 runs sampled) benchmark.js:39:13
Fastest is uuidv1() fill existing array
```