@ctavan 2020-05-03
I should correct myself: I'm happy to merge this if we can ensure that the issue discussed in https://github.com/uuidjs/uuid/pull/267 no longer exists.