@awwit 2020-05-03
@ctavan 

> I should correct myself: I'm happy to merge this if we can ensure that the issue discussed in #267 no longer exists.

This problem still exists. But it can be solved by `.toLowerString()`.